feat: updated TrustDefender to use Signifyd SDKs
