+++ TrustDefender/RLTMXProfiling.xcframework/Info.plist
@@ -20,11 +20,12 @@
 		</dict>
 		<dict>
 			<key>LibraryIdentifier</key>
-			<string>ios-i386_x86_64-simulator</string>
+			<string>ios-arm64_i386_x86_64-simulator</string>
 			<key>LibraryPath</key>
 			<string>RLTMXProfiling.framework</string>
 			<key>SupportedArchitectures</key>
 			<array>
+				<string>arm64</string>
 				<string>i386</string>
 				<string>x86_64</string>
 			</array>
