+++ TrustDefender/RLTMXProfiling.xcframework/ios-arm64_i386_x86_64-simulator/RLTMXProfiling.framework/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module RLTMXProfiling {
+    umbrella header "TMXProfiling.h"
+
+    export *
+    module * { export * }
+}
