--- TrustDefender/RLTMXProfiling.xcframework/ios-i386_x86_64-simulator/RLTMXProfiling.framework/Modules/module.modulemap
@@ -1,6 +1,0 @@
-framework module RLTMXProfiling {
-    umbrella header "TMXProfiling.h"
-
-    export *
-    module * { export * }
-}
