--- TrustDefender/RLTMXProfiling.xcframework/ios-i386_x86_64-simulator/RLTMXProfiling.framework/_CodeSignature/CodeResources
@@ -1,162 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>files</key>
-	<dict>
-		<key>Headers/TMXProfileHandle.h</key>
-		<data>
-		m5Rqa7zaTkhH7BnkJKgzTLvsiew=
-		</data>
-		<key>Headers/TMXProfiling.h</key>
-		<data>
-		Fv5x1SiUYuMeYusnt5raH49S6FM=
-		</data>
-		<key>Headers/TMXStatusCode.h</key>
-		<data>
-		XO0c2Et7sAujX8EQtMygGLH6WTQ=
-		</data>
-		<key>Info.plist</key>
-		<data>
-		RY9bURPLimRGniQqm6Fn3i+8WSM=
-		</data>
-		<key>Modules/module.modulemap</key>
-		<data>
-		OOHZp5MIXv2t7D1clAOp/pmwNGk=
-		</data>
-	</dict>
-	<key>files2</key>
-	<dict>
-		<key>Headers/TMXProfileHandle.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			m5Rqa7zaTkhH7BnkJKgzTLvsiew=
-			</data>
-			<key>hash2</key>
-			<data>
-			W/YteNLgYtybSuWpBKZUehMy922Xi19lcduXX8BfnJQ=
-			</data>
-		</dict>
-		<key>Headers/TMXProfiling.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			Fv5x1SiUYuMeYusnt5raH49S6FM=
-			</data>
-			<key>hash2</key>
-			<data>
-			7ouwrJszNhs9aipHNffwr4vCE7bUhD+iKywUsAjCwYk=
-			</data>
-		</dict>
-		<key>Headers/TMXStatusCode.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			XO0c2Et7sAujX8EQtMygGLH6WTQ=
-			</data>
-			<key>hash2</key>
-			<data>
-			b+9MhNhsocAPSbrab1P1m2qJys5PINMheGKezt6kn9o=
-			</data>
-		</dict>
-		<key>Modules/module.modulemap</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			OOHZp5MIXv2t7D1clAOp/pmwNGk=
-			</data>
-			<key>hash2</key>
-			<data>
-			34VlzFSsXzt0f9IzwNIuvGD8vby2+KCy98Rbu37c5zE=
-			</data>
-		</dict>
-	</dict>
-	<key>rules</key>
-	<dict>
-		<key>^.*</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^Base\.lproj/</key>
-		<dict>
-			<key>weight</key>
-			<real>1010</real>
-		</dict>
-		<key>^version.plist$</key>
-		<true/>
-	</dict>
-	<key>rules2</key>
-	<dict>
-		<key>.*\.dSYM($|/)</key>
-		<dict>
-			<key>weight</key>
-			<real>11</real>
-		</dict>
-		<key>^(.*/)?\.DS_Store$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>2000</real>
-		</dict>
-		<key>^.*</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^Base\.lproj/</key>
-		<dict>
-			<key>weight</key>
-			<real>1010</real>
-		</dict>
-		<key>^Info\.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^PkgInfo$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^embedded\.provisionprofile$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^version\.plist$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-	</dict>
-</dict>
-</plist>
