+++ TrustDefender/RLTMXProfilingConnections.xcframework/ios-arm64_i386_x86_64-simulator/RLTMXProfilingConnections.framework/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module RLTMXProfilingConnections {
+    umbrella header "TMXProfilingConnections.h"
+
+    export *
+    module * { export * }
+}
