+++ TrustDefender/RLTMXProfilingConnections.xcframework/ios-arm64_i386_x86_64-simulator/RLTMXProfilingConnections.framework/_CodeSignature/CodeResources
@@ -1,0 +1,147 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>files</key>
+	<dict>
+		<key>Headers/TMXProfilingConnections.h</key>
+		<data>
+		IwIWvHm+9Q+yI49PJ00L7SeNlkE=
+		</data>
+		<key>Headers/TMXProfilingConnectionsProtocol.h</key>
+		<data>
+		6wDxVmlLMkbaYF7oOijLYxRGMu0=
+		</data>
+		<key>Info.plist</key>
+		<data>
+		Qrk/42EqMef7jVZMhZUKl1yLCJ8=
+		</data>
+		<key>Modules/module.modulemap</key>
+		<data>
+		YBWsTFEF7pD/jlcPSu3L0g1IxoA=
+		</data>
+	</dict>
+	<key>files2</key>
+	<dict>
+		<key>Headers/TMXProfilingConnections.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			IwIWvHm+9Q+yI49PJ00L7SeNlkE=
+			</data>
+			<key>hash2</key>
+			<data>
+			A8TimCl+4+rCIB33KvgkI8G2pM5UYpmApASl6qrsHus=
+			</data>
+		</dict>
+		<key>Headers/TMXProfilingConnectionsProtocol.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			6wDxVmlLMkbaYF7oOijLYxRGMu0=
+			</data>
+			<key>hash2</key>
+			<data>
+			HfhrPd20XcFjOcxi96bL9Dox80OxtwOZM1B+oOwwnQk=
+			</data>
+		</dict>
+		<key>Modules/module.modulemap</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			YBWsTFEF7pD/jlcPSu3L0g1IxoA=
+			</data>
+			<key>hash2</key>
+			<data>
+			NDIWkDlNlwJgiFbns1aWbNdbDIaiVLq7hTjRy8XiKcI=
+			</data>
+		</dict>
+	</dict>
+	<key>rules</key>
+	<dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^version.plist$</key>
+		<true/>
+	</dict>
+	<key>rules2</key>
+	<dict>
+		<key>.*\.dSYM($|/)</key>
+		<dict>
+			<key>weight</key>
+			<real>11</real>
+		</dict>
+		<key>^(.*/)?\.DS_Store$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>2000</real>
+		</dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^Info\.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^PkgInfo$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^embedded\.provisionprofile$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^version\.plist$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+	</dict>
+</dict>
+</plist>
