--- TrustDefender/RLTMXProfilingConnections.xcframework/ios-i386_x86_64-simulator/RLTMXProfilingConnections.framework/Modules/module.modulemap
@@ -1,6 +1,0 @@
-framework module RLTMXProfilingConnections {
-    umbrella header "TMXProfilingConnections.h"
-
-    export *
-    module * { export * }
-}
